evaluation/requirements.txt
torch
torchvision
pillow
numpy
diffusers
transformers
accelerate
safetensors
fastapi
uvicorn[standard]
python-multipart

# Evaluation dependencies
torchmetrics
scipy
tqdm
ftfy
regex
matplotlib
git+https://github.com/openai/CLIP.git